Refractor, small changes
