--- PrincessPaperplane/configs/aliases.py
@@ -1,14 +1,0 @@
-
-"""
-Aliases:
-	<BASE_COMMAND>:
-		<NAME> : [aliases]
-"""
-
-RANK = {
-	'rank' : ['rank', 'rang', 'level', 'lvl']
-}
-
-DICE = {
-	'dice': ['dice', 'w', 'd']
-}
