+++ PrincessPaperplane/configs/cmd_config.py
@@ -1,0 +1,6 @@
+
+
+ALIASES = {
+	"rank": ['rank', 'rang', 'level', 'lvl'],
+	"dice": ['dice', 'w', 'd']
+}
