+++ PrincessPaperplane/configs/xp_config.py
@@ -1,2 +1,7 @@
+"""Settings for XP
+"""
+
 BASE = 15
-RANDOM_RANGE = (0,10)
+RANDOM_RANGE = (0, 10)
+
+COOLDOWN_SEC = 60
