+++ PrincessPaperplane/utility/checks.py
@@ -1,5 +1,7 @@
-from discord.ext import commands
 from typing import Optional
+
+from discord.ext import commands
+
 
 def is_in_channel(channel_id : int, check_on_server_id : Optional[int] = 0):
     """Used as decorator that checks if bot can post in this channel"""
